judging folder refactor
