+++ src/containers/JudgingPanel.js
@@ -4,8 +4,8 @@ import { H1, H3, P, A } from '../components/Typography'
 import { Card } from '../components/Common'
 import Accordion from '../components/Accordion'
 import { syncToFirebase, projectsRef } from '../utility/firebase'
-import ProjectTable from '../components/Judging/ProjectTable'
-import SponsorSubmissions from '../components/Judging/SponsorSubmissions'
+import ProjectTable from '../components/Judging/Admin/ProjectTable'
+import SponsorSubmissions from '../components/Judging/Admin/SponsorSubmissions'
 import { MoonLoader } from 'react-spinners'
 import ProgressBar from '../components/ProgressBar'
