+++ src/pages/Judging/index.js
@@ -4,7 +4,7 @@ import { Link } from 'wouter'
 import HeroPage, { Loading, JudgingNotOpen } from '../../components/HeroPage'
 import { db, firestore, getLivesiteDoc, applicantsRef, projectsRef } from '../../utility/firebase'
 import { formatProject } from '../../utility/utilities'
-import JudgingCard from '../../components/JudgingCard'
+import JudgingCard from '../../components/Judging/JudgingCard'
 import { A } from '../../components/Typography'
 import { useAuth } from '../../utility/Auth'
